# decosjoin/openapi.yml
openapi: 3.0.3
info:
  title: Decos join koppel api API
  description: Dit is het koppelstuk tussen mijn amsterdam en decos join
  version: 1.0.0
paths:
  /decosjoin/getvergunningen:
    parameters:
      - name: x-saml-attribute-token1
        in: header
        description: TMA encoded BSN
        required: true
        schema:
          type: string
    get:
      summary: Returns a list verguningen.
      description: Returns a list of vergunningen
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: "#/components/schemas/Zaken"
                  status:
                    type: string
                    enum:
                      - "OK"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /decosjoin/listdocuments/{encrypted_zaak_id}:
    parameters:
      - name: x-saml-attribute-token1
        in: header
        description: TMA encoded BSN
        required: true
        schema:
          type: string
      - name: encrypted_zaak_id
        in: path
        description: fernet encoded string of zaak id
        schema:
          type: string
        required: true
    get:
      summary: Returns a list of documents for zaak
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
                  status:
                    type: string
                    enum:
                      - "OK"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /decosjoin/document/{encrypted_doc_id}:
    parameters:
      - name: x-saml-attribute-token1
        in: header
        description: TMA encoded BSN
        required: true
        schema:
          type: string
      - name: encrypted_doc_id
        in: path
        description: fernet encoded string of document id
        schema:
          type: string
        required: true
    get:
      summary: Returns a document as http download
      responses:
        "200":
          description: Returns document blob
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /status/health:
    get:
      summary: returns "OK"
      responses:
        "200":
          description: Returns "OK"
          content:
            application/json:
              schema:
                type: string
                enum:
                  - "OK"
components:
  schemas:
    Zaken:
      type: array
      items:
        anyOf:
          - $ref: "#/components/schemas/Zaak_TVM_RVV_Object"
          - $ref: "#/components/schemas/Zaak_VakantieVerhuurVergunning"
          - $ref: "#/components/schemas/Zaak_VakantieVerhuur"
          - $ref: "#/components/schemas/Zaak_VakantieVerhuurAfmelding"
          - $ref: "#/components/schemas/Zaak_BBVergunning"
          - $ref: "#/components/schemas/Zaak_GPP"
          - $ref: "#/components/schemas/Zaak_GPK"
          - $ref: "#/components/schemas/Zaak_EvenementMelding"
          - $ref: "#/components/schemas/Zaak_EvenementVergunning"
          - $ref: "#/components/schemas/Zaak_Omzettingsvergunning"
          - $ref: "#/components/schemas/Zaak_ERVV_TVM"
          - $ref: "#/components/schemas/Zaak_BZP"
          - $ref: "#/components/schemas/Zaak_BZB"
    Zaak:
      type: object
      properties:
        title:
          type: string
          example: "SB RVV ontheffing hele stad"
        caseType:
          type: string
          example: "TVM - RVV - Object"
        identifier:
          type: string
          example: "Z/20/1234567"
        dateRequest:
          $ref: "#/components/schemas/Date"
        status:
          type: string
          example: "Ontvangen"
        decision:
          type: string
          example: "?"
          nullable: true
        dateDecision:
          $ref: "#/components/schemas/Date"
        documentsUrl:
          type: string
          example: "/api/decos/listdocuments/<fernetstring>"
    Zaak_TVM_RVV_Object:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: TVM - RVV - Object
              enum:
                - TVM - RVV - Object
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            kenteken:
              $ref: "#/components/schemas/Kenteken"
            location:
              $ref: "#/components/schemas/Location"
            timeEnd:
              $ref: "#/components/schemas/Time"
            timeStart:
              $ref: "#/components/schemas/Time"
    Zaak_VakantieVerhuurVergunning:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Vakantieverhuur vergunningsaanvraag
              enum:
                - Vakantieverhuur vergunningsaanvraag
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
    Zaak_VakantieVerhuur:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Vakantieverhuur
              enum:
                - Vakantieverhuur
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            location:
              $ref: "#/components/schemas/Location"
    Zaak_VakantieVerhuurAfmelding:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Vakantieverhuur afmelding
              enum:
                - Vakantieverhuur afmelding
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
    Zaak_BBVergunning:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: B&B - vergunning
              enum:
                - B&B - vergunning
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            dateWorkflowActive:
              $ref: "#/components/schemas/Date"
            requester:
              type: string
              nullable: true
            owner:
              type: string
              nullable: true
            hasTransitionAgreement:
              type: boolean
    Zaak_GPP:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: GPP
              enum:
                - GPP
            kenteken:
              $ref: "#/components/schemas/Kenteken"
            location:
              $ref: "#/components/schemas/Location"
    Zaak_GPK:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: GPK
              enum:
                - GPK
            dateEnd:
              $ref: "#/components/schemas/Date"
            cardNumber:
              type: string
              nullable: true
            cardType:
              type: string
              nullable: true
    Zaak_EvenementMelding:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Evenement melding
              enum:
                - Evenement melding
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            location:
              $ref: "#/components/schemas/Location"
            timeEnd:
              $ref: "#/components/schemas/Time"
            timeStart:
              $ref: "#/components/schemas/Time"
    Zaak_EvenementVergunning:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Evenement vergunning
              enum:
                - Evenement vergunning
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            location:
              $ref: "#/components/schemas/Location"
            timeEnd:
              $ref: "#/components/schemas/Time"
            timeStart:
              $ref: "#/components/schemas/Time"
    Zaak_Omzettingsvergunning:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Omzettingsvergunning
              enum:
                - Omzettingsvergunning
            location:
              $ref: "#/components/schemas/Location"
            dateWorkflowActive:
              $ref: "#/components/schemas/Date"
    Zaak_ERVV_TVM:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: E-RVV - TVM
              enum:
                - E-RVV - TVM
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            location:
              $ref: "#/components/schemas/Location"
            timeEnd:
              $ref: "#/components/schemas/Time"
            timeStart:
              $ref: "#/components/schemas/Time"
    Zaak_BZP:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Parkeerontheffingen Blauwe zone particulieren
              enum:
                - Parkeerontheffingen Blauwe zone particulieren
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            kenteken:
              $ref: "#/components/schemas/Kenteken"
    Zaak_BZB:
      allOf:
        - $ref: "#/components/schemas/Zaak"
        - type: object
          properties:
            caseType:
              type: string
              example: Parkeerontheffingen Blauwe zone bedrijven
              enum:
                - Parkeerontheffingen Blauwe zone bedrijven
            dateEnd:
              $ref: "#/components/schemas/Date"
            dateStart:
              $ref: "#/components/schemas/Date"
            kenteken:
              $ref: "#/components/schemas/Kenteken"
    Document:
      type: object
      properties:
        id:
          type: string
          example: D/2296393
        title:
          type: string
          example: Besluit toekenning
        url:
          type: string
          example: "api/decosjoin/document/<fernetstring>"
    Date:
      type: string
      format: date
      example: "2020-06-16"
      nullable: true
    Time:
      type: string
      example: "10:15"
      nullable: true
    DateTime:
      type: string
      format: date-time
      example: "2020-06-08T00:00:00"
      nullable: true
    Kenteken:
      type: string
      example: "aa-11-aa"
      nullable: true
    Location:
      type: string
      nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - "ERROR"
        message:
          type: string